orchestrator_real/requirements.txt
google-adk>=0.1.0
a2a-python>=0.1.0
google-cloud-aiplatform>=1.38.0
httpx>=0.25.0
uvicorn>=0.24.0
fastapi>=0.104.0
